// src/screens/PaymentPage/PaySection/YandexPay/yapay.types.ts
type Listener<T> = (event: T) => void;
type UnsubscribeCallback = () => void;

interface TypedObject<T> {
  type: T;
}

/**
 * Общие типы.
 */

/**
 * Должно быть больше 0 и не содержать больше двух знаков после запятой.
 * Например: 1.12, 5.1, 10.
 */
export type Price = string;

/**
 * Платежная криптограмма.
 */
export type PaymentToken = string;
// ...
export interface OrderTotal {
  /**
   * Текст для поля total.
   */
  label?: string;
  amount: Price;
}

export interface OrderItem {
  label: string;
  amount: Price;
}

export interface Order {
  /**
   * ID заказа на стороне продавца.
   */
  id: string;
  total: OrderTotal;
  items?: OrderItem[];
}
// ...
export type PaymentMethodInfo = CardPaymentMethodInfo;

/**
 * Запрашиваемые контактные данные плательщика.
 */
export interface RequiredBillingContactFields {
  email?: boolean;
}

/**
 * Запрашиваемые данные с платежной формы.
 */
export interface RequiredFields {
  billingContact?: RequiredBillingContactFields;
}

/**
 * Возвращаемые данные плательщика.
 */
export interface BillingContactField {
  // Поле появится, если было запрошено в RequiredBillingContactFields.
  email?: string;
}
// ...
export interface UpdatePaymentData {
  order?: Order;
}

/**
 * События формы.
 */

export enum PaymentEventType {
  Ready = 'ready',
  Abort = 'abort',
  Error = 'error',
  Process = 'process',
  Change = 'change',
}

export type ReadyEvent = TypedObject<PaymentEventType.Ready>;

export enum AbortEventReason {
  /**
   * Покупатель закрыл форму оплаты.
   */
  Close = 'CLOSE',

  /**
   * Покупатель не успел совершить оплату.
   */
  Timeout = 'TIMEOUT',
}

export interface AbortEvent extends TypedObject<PaymentEventType.Abort> {
  reason?: AbortEventReason;
}

export interface ErrorEvent extends TypedObject<PaymentEventType.Error> {
  reason?: string;
  details?: any;
}

export interface ProcessEvent extends TypedObject<PaymentEventType.Process> {
  token: PaymentToken;
  paymentMethodInfo: PaymentMethodInfo;
  // Поле появится, если было запрошено в PaymentSheet.
  billingContact?: BillingContactField;
}

export type PaymentEvent = ReadyEvent | AbortEvent | ErrorEvent | ProcessEvent;

/**
 * Сообщения SDK.
 */

export enum MessageType {
  Payment = 'payment',
  Complete = 'complete',
}

export interface PaymentMessage extends TypedObject<MessageType.Payment> {
  sheet: PaymentSheet;
}

export enum CompleteReason {
  Success = 'success',
  Error = 'error',
  Close = 'close',
  AuthRedirect = 'auth-redirect',
}

export interface CompleteMessage extends TypedObject<MessageType.Complete> {
  reason: CompleteReason;
  errors?: any;
}

export type Message = PaymentMessage | CompleteMessage;

/**
 * Настройки платежной кнопки.
 */

export type ButtonParent = HTMLElement | ShadowRoot;

export enum ButtonType {
  Simple = 'SIMPLE',
  Pay = 'PAY',
}

export enum ButtonTheme {
  White = 'WHITE',
  WhiteOutlined = 'WHITE-OUTLINED',
  Black = 'BLACK',
}

export enum ButtonWidth {
  Auto = 'AUTO',
  Max = 'MAX',
}

export interface ButtonOptions {
  /**
   * @default ButtonType.SIMPLE
   */
  type?: ButtonType;

  /**
   * @default ButtonTheme.RED
   */
  theme?: ButtonTheme;

  /**
   * @default ButtonWidth.AUTO
   */
  width?: ButtonWidth;
}

/**
 * События платежной кнопки.
 */

export enum ButtonEventType {
  Click = 'CLICK',
}

/**
 * Конструктор платежной кнопки.
 */

export declare class Button {
  constructor(options: ButtonOptions);

  static create(options: ButtonOptions): Button;

  /**
   * Добавить кнопку в DOM-дерева.
   */
  mount(parent: ButtonParent): void;

  /**
   * Удалить кнопку из DOM-дерева.
   */
  unmount(): void;

  /**
   * Удаляет все слушатели с кнопки и кнопку из DOM-дерева.
   */
  destroy(): void;

  /**
   * Слушатель на клик.
   * Используйте этот параметр для запуска процесса оплаты на клик по кнопке.
   */
  on(type: ButtonEventType.Click, listener: Listener<MouseEvent>): void;
}

/**
 * Платеж.
 */

export interface Payment {
  /**
   * Создает привязанный к платежу экземпляр кнопки.
   */
  createButton(options: ButtonOptions): Button;

  /**
   * Обновляет данные платежа.
   */
  update(updateData: UpdatePaymentData): void;

  /**
   * Запускает процесс оплаты.
   */
  checkout(): void;

  /**
   * Удаляет платеж.
   */
  destroy(): void;

  /**
   * Подписывает на события формы платежа.
   */

  /**
   * Форма оплаты загружена.
   */
  on(type: PaymentEventType.Ready, listener: Listener<ReadyEvent>): UnsubscribeCallback;

  /**
   * Оплата была отменена. Пример: пользователь закрыл форму.
   */
  on(type: PaymentEventType.Abort, listener: Listener<AbortEvent>): UnsubscribeCallback;

  /**
   * Произошла ошибка во время оплаты.
   */
  on(type: PaymentEventType.Error, listener: Listener<ErrorEvent>): UnsubscribeCallback;

  /**
   * Запустился процесс оплаты.
   * Вместе с событием возвращается платежный токен.
   */
  on(type: PaymentEventType.Process, listener: Listener<ProcessEvent>): UnsubscribeCallback;

  /**
   * Дополнительное управление состоянием формы оплаты.
   */

  /**
   * Сообщить форме об успешной оплате.
   */
  complete(reason: CompleteReason.Success): void;

  /**
   * Сообщить форме об ошибке оплаты.
   */
  complete(reason: CompleteReason.Error, errors: unknown): void;

  /**
   * Закрыть платежную форму.
   */
  complete(reason: CompleteReason.Close): void;

  /**
   * Сообщить форме о редиректе покупателя на стороннюю форму оплаты.
   */
  complete(reason: CompleteReason.AuthRedirect): void;
}